+++ tests/test_logger.py
@@ -6,6 +6,7 @@ class RootLoggerTest(unittest.TestCase):
     def test_info(self):
         with self.assertLogs(root_logger, level='INFO') as cm:
             root_logger.info("Test message")
+            self.assertEqual(cm.records[0].getMessage(), "Test message")
         # Assert that the log message was written to the console correctly
         # You can use a library like `pytest` for more advanced assertions
